// src/modules/invoice/usecase/find-all-invoice/find-all-invoice.dto.ts
export interface FindAllInvoiceUseCaseOutputDto {
  invoices: {
    id: string
    name: string
    document: string
    address: {
      street: string
      number: string
      complement: string
      city: string
      state: string
      zipCode: string
    }
    items: {
      id: string
      name: string
      price: number
    }[]
    total: number
    createdAt: Date
  }[]
}